Start working on simple collision detection
A quick and dirty implementation of sequential collision detection for
player shots. Sequential checking should be simple and effective.
Functions need to be made more general so that they can be used for
enemy shots and bunker defenses. Collision checking needs to have a
limit on movement speed and use that limit to move towards the
destination while checking for collisions. This should prevent
tunneling of fast moving objects.