Record demo as binary and overwrite existing file
The demo is now stored as a binary file. This change also stores the
initial random seed for the game. The demo file is now written at the
end of the game, instead of appending to file each frame. This also
fixes the issue of appending data to existing file, as the file is now
overwritten.